1937 문제풀이

+++ src/Java/[10000-19999]/[10000-10999]/10986/Main.java
@@ -6,7 +6,7 @@ public class Main {
     static int m;
     static long[] mod;
     static long[] num;
-    static long ans;
+    static int ans;
 
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
